feat: Run the backstage upgrade.
