+++ .gitignore
@@ -7,7 +7,6 @@ logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-lerna-debug.log*
 
 # Coverage directory generated when running tests with coverage
 coverage
@@ -49,3 +48,6 @@ site
 
 # E2E test reports
 e2e-test-report/
+
+# Cache
+.cache/
