+++ backstage.json
@@ -1,3 +1,3 @@
 {
-  "version": "1.36.1"
+  "version": "1.49.2"
 }
