+++ examples/example-docs/docs/index.md
@@ -1,0 +1,5 @@
+# Example TechDocs page
+
+This is an example TechDocs page!
+
+The source of this documentation for the `example-website` component is found in the `examples/example-docs` directory, which is referenced from `examples/entities.yaml`. Typically the `docs` directory would live alongside the `catalog-info.yaml` file of a component instead, but in this case it is separated out to avoid cluttering the examples directory.
