+++ packages/app/package.json
@@ -16,38 +16,36 @@
   "dependencies": {
     "@backstage-community/plugin-github-actions": "^0.6.19",
     "@backstage-community/plugin-tech-radar": "^0.7.6",
-    "@backstage/app-defaults": "^1.5.17",
-    "@backstage/catalog-model": "^1.7.3",
-    "@backstage/cli": "^0.30.0",
-    "@backstage/core-app-api": "^1.15.5",
-    "@backstage/core-components": "^0.16.4",
-    "@backstage/core-plugin-api": "^1.10.4",
-    "@backstage/integration-react": "^1.2.4",
-    "@backstage/plugin-api-docs": "^0.12.4",
-    "@backstage/plugin-catalog": "^1.27.0",
-    "@backstage/plugin-catalog-common": "^1.1.3",
-    "@backstage/plugin-catalog-graph": "^0.4.16",
-    "@backstage/plugin-catalog-import": "^0.12.10",
-    "@backstage/plugin-catalog-react": "^1.15.2",
-    "@backstage/plugin-org": "^0.6.36",
-    "@backstage/plugin-permission-react": "^0.4.31",
-    "@backstage/plugin-scaffolder": "^1.28.0",
-    "@backstage/plugin-search": "^1.4.23",
-    "@backstage/plugin-search-react": "^1.8.6",
-    "@backstage/plugin-techdocs": "^1.12.3",
-    "@backstage/plugin-techdocs-module-addons-contrib": "^1.1.21",
-    "@backstage/plugin-techdocs-react": "^1.2.14",
-    "@backstage/plugin-user-settings": "^0.8.19",
-    "@backstage/theme": "^0.6.4",
+    "@backstage/cli": "^0.36.0",
+    "@backstage/core-components": "^0.18.8",
+    "@backstage/core-plugin-api": "^1.12.4",
+    "@backstage/frontend-defaults": "^0.5.0",
+    "@backstage/frontend-plugin-api": "^0.15.1",
+    "@backstage/integration-react": "^1.2.16",
+    "@backstage/plugin-api-docs": "^0.13.5",
+    "@backstage/plugin-app-react": "^0.2.1",
+    "@backstage/plugin-app-visualizer": "^0.2.1",
+    "@backstage/plugin-catalog": "^2.0.1",
+    "@backstage/plugin-catalog-graph": "^0.6.0",
+    "@backstage/plugin-catalog-import": "^0.13.11",
+    "@backstage/plugin-notifications": "^0.5.15",
+    "@backstage/plugin-org": "^0.7.0",
+    "@backstage/plugin-scaffolder": "^1.36.1",
+    "@backstage/plugin-search": "^1.7.0",
+    "@backstage/plugin-signals": "^0.0.29",
+    "@backstage/plugin-techdocs": "^1.17.2",
+    "@backstage/plugin-techdocs-module-addons-contrib": "^1.1.34",
+    "@backstage/plugin-user-settings": "^0.9.1",
+    "@backstage/ui": "^0.13.1",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "react": "18.1.0",
     "react-dom": "18.1.0",
-    "react-router": "^6.3.0",
-    "react-router-dom": "^6.3.0"
+    "react-router": "^6.30.2",
+    "react-router-dom": "^6.30.2"
   },
   "devDependencies": {
-    "@backstage/test-utils": "^1.7.5",
+    "@backstage/frontend-test-utils": "^0.5.1",
     "@playwright/test": "^1.32.3",
     "@testing-library/dom": "^9.0.0",
     "@testing-library/jest-dom": "^6.0.0",
