+++ packages/app/src/modules/nav/LogoIcon.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { makeStyles } from '@material-ui/core';
 
 const useStyles = makeStyles({
@@ -11,7 +10,7 @@ const useStyles = makeStyles({
   },
 });
 
-const LogoIcon = () => {
+export const LogoIcon = () => {
   const classes = useStyles();
 
   return (
@@ -27,5 +26,3 @@ const LogoIcon = () => {
     </svg>
   );
 };
-
-export default LogoIcon;
