+++ packages/app/src/modules/nav/SidebarLogo.tsx
@@ -1,0 +1,36 @@
+import {
+  Link,
+  sidebarConfig,
+  useSidebarOpenState,
+} from '@backstage/core-components';
+import { makeStyles } from '@material-ui/core';
+import { LogoFull } from './LogoFull';
+import { LogoIcon } from './LogoIcon';
+
+const useSidebarLogoStyles = makeStyles({
+  root: {
+    width: sidebarConfig.drawerWidthClosed,
+    height: 3 * sidebarConfig.logoHeight,
+    display: 'flex',
+    flexFlow: 'row nowrap',
+    alignItems: 'center',
+    marginBottom: -14,
+  },
+  link: {
+    width: sidebarConfig.drawerWidthClosed,
+    marginLeft: 24,
+  },
+});
+
+export const SidebarLogo = () => {
+  const classes = useSidebarLogoStyles();
+  const { isOpen } = useSidebarOpenState();
+
+  return (
+    <div className={classes.root}>
+      <Link to="/" underline="none" className={classes.link} aria-label="Home">
+        {isOpen ? <LogoFull /> : <LogoIcon />}
+      </Link>
+    </div>
+  );
+};
