+++ packages/backend/package.json
@@ -16,39 +16,43 @@
     "build-image": "docker build ../.. -f Dockerfile --tag backstage"
   },
   "dependencies": {
-    "@backstage/backend-defaults": "^0.8.1",
-    "@backstage/config": "^1.3.2",
-    "@backstage/integration": "^1.16.1",
-    "@backstage/plugin-app-backend": "^0.4.5",
-    "@backstage/plugin-auth-backend": "^0.24.3",
-    "@backstage/plugin-auth-backend-module-github-provider": "^0.3.0",
-    "@backstage/plugin-auth-backend-module-guest-provider": "^0.2.5",
-    "@backstage/plugin-auth-node": "^0.6.0",
-    "@backstage/plugin-catalog-backend": "^1.31.0",
-    "@backstage/plugin-catalog-backend-module-github": "^0.7.10",
-    "@backstage/plugin-catalog-backend-module-github-org": "^0.3.7",
-    "@backstage/plugin-catalog-backend-module-logs": "^0.1.7",
-    "@backstage/plugin-catalog-backend-module-scaffolder-entity-model": "^0.2.5",
-    "@backstage/plugin-events-backend-module-github": "^0.2.17",
-    "@backstage/plugin-permission-backend": "^0.5.54",
-    "@backstage/plugin-permission-backend-module-allow-all-policy": "^0.2.5",
-    "@backstage/plugin-permission-common": "^0.8.4",
-    "@backstage/plugin-permission-node": "^0.8.8",
-    "@backstage/plugin-proxy-backend": "^0.5.11",
-    "@backstage/plugin-scaffolder-backend": "^1.30.0",
-    "@backstage/plugin-search-backend": "^1.8.2",
-    "@backstage/plugin-search-backend-module-catalog": "^0.3.1",
-    "@backstage/plugin-search-backend-module-pg": "^0.5.41",
-    "@backstage/plugin-search-backend-module-techdocs": "^0.3.6",
-    "@backstage/plugin-search-backend-node": "^1.3.8",
-    "@backstage/plugin-techdocs-backend": "^1.11.6",
+    "@backstage/backend-defaults": "^0.16.0",
+    "@backstage/config": "^1.3.6",
+    "@backstage/integration": "^2.0.0",
+    "@backstage/plugin-app-backend": "^0.5.12",
+    "@backstage/plugin-auth-backend": "^0.27.3",
+    "@backstage/plugin-auth-backend-module-github-provider": "^0.5.1",
+    "@backstage/plugin-auth-backend-module-guest-provider": "^0.2.17",
+    "@backstage/plugin-auth-node": "^0.6.14",
+    "@backstage/plugin-catalog-backend": "^3.5.0",
+    "@backstage/plugin-catalog-backend-module-github": "^0.13.0",
+    "@backstage/plugin-catalog-backend-module-github-org": "^0.3.20",
+    "@backstage/plugin-catalog-backend-module-logs": "^0.1.20",
+    "@backstage/plugin-catalog-backend-module-scaffolder-entity-model": "^0.2.18",
+    "@backstage/plugin-events-backend-module-github": "^0.4.10",
+    "@backstage/plugin-notifications-backend": "^0.6.3",
+    "@backstage/plugin-permission-backend": "^0.7.10",
+    "@backstage/plugin-permission-backend-module-allow-all-policy": "^0.2.17",
+    "@backstage/plugin-permission-common": "^0.9.7",
+    "@backstage/plugin-permission-node": "^0.10.11",
+    "@backstage/plugin-proxy-backend": "^0.6.11",
+    "@backstage/plugin-scaffolder-backend": "^3.2.0",
+    "@backstage/plugin-scaffolder-backend-module-github": "^0.9.7",
+    "@backstage/plugin-scaffolder-backend-module-notifications": "^0.1.20",
+    "@backstage/plugin-search-backend": "^2.1.0",
+    "@backstage/plugin-search-backend-module-catalog": "^0.3.13",
+    "@backstage/plugin-search-backend-module-pg": "^0.5.53",
+    "@backstage/plugin-search-backend-module-techdocs": "^0.4.12",
+    "@backstage/plugin-search-backend-node": "^1.4.2",
+    "@backstage/plugin-signals-backend": "^0.3.13",
+    "@backstage/plugin-techdocs-backend": "^2.1.6",
     "app": "link:../app",
     "node-gyp": "^10.0.0",
     "pg": "^8.11.3"
   },
   "devDependencies": {
-    "@backstage/cli": "^0.30.0",
-    "better-sqlite3": "^11.8.1"
+    "@backstage/cli": "^0.36.0",
+    "better-sqlite3": "^12.0.0"
   },
   "files": [
     "dist"
