+++ packages/backend/src/index.ts
@@ -7,14 +7,21 @@
  */
 
 import { createBackend } from '@backstage/backend-defaults';
-import { eventsModuleGithubEventRouter } from '@backstage/plugin-events-backend-module-github/alpha';
+import eventsModuleGithubEventRouter from '@backstage/plugin-events-backend-module-github';
 
 const backend = createBackend();
 
 backend.add(import('@backstage/plugin-app-backend'));
 backend.add(import('@backstage/plugin-proxy-backend'));
+
+// scaffolder plugin
 backend.add(import('@backstage/plugin-scaffolder-backend'));
 backend.add(import('@backstage/plugin-scaffolder-backend-module-github'));
+backend.add(
+  import('@backstage/plugin-scaffolder-backend-module-notifications'),
+);
+
+// techdocs plugin
 backend.add(import('@backstage/plugin-techdocs-backend'));
 
 // auth plugin
@@ -51,5 +58,9 @@ backend.add(import('@backstage/plugin-search-backend-module-pg'));
 // search collators
 backend.add(import('@backstage/plugin-search-backend-module-catalog'));
 backend.add(import('@backstage/plugin-search-backend-module-techdocs'));
+
+// notifications and signals plugins
+backend.add(import('@backstage/plugin-notifications-backend'));
+backend.add(import('@backstage/plugin-signals-backend'));
 
 backend.start();
