+++ playwright.config.ts
@@ -28,8 +28,14 @@ export default defineConfig({
     ? []
     : [
         {
-          command: 'yarn dev',
+          command: 'yarn start app',
           port: 3000,
+          reuseExistingServer: true,
+          timeout: 60_000,
+        },
+        {
+          command: 'yarn start backend',
+          port: 7007,
           reuseExistingServer: true,
           timeout: 60_000,
         },
